refactor(update-collection-v3): make it easier to add new migrations
A migration should be a stand-alone function
that takes a string as input (the contents of the values.yaml file)
and gives a string on output (contents of migrated values.yaml file).

This change extracts the existing migrations
for kube-prometheus-stack and events
into one migration (as they are tightly coupled)
and prepares infrastructure for easy adding of new migrations.